requirements.txt

streamlit
pandas
openpyxl
google-generativeai
python-dotenv
geopy
pydeck
httpx
requests
sentence-transformers
faiss-cpu
langchain
langchain-community
langchain-google-genai
ollama
shapely
